Fix for #1085 - modules not correctly being disabled at the end of a game session. Moved CrashReporter to the PC facade. Added missing 'meta' subproject to settings.gradle
