+++ facades/PC/src/main/java/org/terasology/engine/Terasology.java
@@ -32,7 +32,6 @@ import org.terasology.engine.subsystem.lwjgl.LwjglAudio;
 import org.terasology.engine.subsystem.lwjgl.LwjglGraphics;
 import org.terasology.engine.subsystem.lwjgl.LwjglInput;
 import org.terasology.engine.subsystem.lwjgl.LwjglTimer;
-import org.terasology.utilities.CrashReporter;
 
 import com.google.common.collect.Lists;
