+++ settings.gradle
@@ -2,7 +2,7 @@ import groovy.io.FileType
 
 rootProject.name = 'Terasology'
 
-include 'engine', 'engine-tests', 'facades', 'libs', 'modules'
+include 'engine', 'engine-tests', 'facades', 'meta', 'libs', 'modules'
 
 // Handy little snippet found online that'll "fake" having nested settings.gradle files under /modules, /libs, etc
 rootDir.eachDir { possibleSubprojectDir ->
